okr-agent/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
markdown-it-py
python-frontmatter
langchain
langchain-community
